Rename route for user and team registration forms
